fix migration with correcte column type

+++ database/migrations/2019_05_29_114336_create_role_user_table.php
@@ -15,8 +15,8 @@ class CreateRoleUserTable extends Migration
     {
         Schema::create('role_user', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigIncrements('role_id');
-            $table->bigIncrements('user_id');
+            $table->unsignedInteger('role_id');
+            $table->unsignedInteger('user_id');
             $table->timestamps();
         });
     }
